+++ src/main/java/cart/auth/BasicAuthorizationExtractor.java
@@ -6,6 +6,7 @@ import org.apache.tomcat.util.codec.binary.Base64;
 import javax.servlet.http.HttpServletRequest;
 
 public class BasicAuthorizationExtractor {
+
     private static final String BASIC_TYPE = "Basic";
     String AUTHORIZATION = "Authorization";
     private static final String DELIMITER = ":";
