+++ src/main/java/cart/business/CartService.java
@@ -1,16 +1,11 @@
 package cart.business;
 
-import cart.entity.Cart;
-import cart.entity.Member;
 import cart.entity.Product;
 import cart.persistence.CartDao;
-import cart.presentation.dto.CartRequest;
-import cart.presentation.dto.CartResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CartService {
@@ -21,13 +16,6 @@ public class CartService {
         this.cartDao = cartDao;
     }
 
-//    @Transactional
-//    public Integer create(CartRequest request) {
-//        Cart cart = makeCartFromRequest(request);
-//
-//        return cartDao.insert(cart);
-//    }
-
     public List<Product> findProductsByMemberId(Integer memberId) {
         return cartDao.findAllProductsByMemberId(memberId);
     }
@@ -35,17 +23,5 @@ public class CartService {
     @Transactional
     public Integer delete(Integer id) {
         return cartDao.remove(id);
-    }
-
-    public Optional<Member> findMemberByEmail(String email) {
-        return cartDao.findByEmail(email);
-    }
-
-    private Cart makeCartFromRequest(CartRequest request) {
-        return new Cart(null, request.getMemberId());
-    }
-
-    private Cart makeCartFromResponse(CartResponse response) {
-        return new Cart(response.getId(), response.getMemberId());
     }
 }
