+++ src/main/java/cart/business/MemberService.java
@@ -2,12 +2,11 @@ package cart.business;
 
 import cart.entity.Member;
 import cart.persistence.MemberDao;
-import cart.presentation.dto.MemberRequest;
-import cart.presentation.dto.MemberResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MemberService {
@@ -18,30 +17,17 @@ public class MemberService {
         this.memberDao = memberDao;
     }
 
-    @Transactional
-    public Integer create(MemberRequest request) {
-        Member member = makeMemberFromRequest(request);
-        memberDao.findSameMemberExist(member);
-
-        return memberDao.insert(member);
-    }
-
     @Transactional(readOnly = true)
     public List<Member> read() {
         return memberDao.findAll();
     }
 
+    public Optional<Member> findMemberByEmail(String email) {
+        return memberDao.findByEmail(email);
+    }
+
     @Transactional
     public Integer delete(Integer id) {
         return memberDao.remove(id);
-    }
-
-
-    private Member makeMemberFromRequest(MemberRequest request) {
-        return new Member(null, request.getEmail(), request.getPassword());
-    }
-
-    private Member makeMemberFromResponse(MemberResponse response) {
-        return new Member(response.getId(), response.getEmail(), response.getPassword());
     }
 }
