+++ src/main/resources/templates/cart.html
@@ -64,12 +64,6 @@
 
     });
 
-
-
-
-
-
-
 </script>
 </body>
 </html>
